Unify editor UI and add wiki sidebar
